+++ test/integration/default/default_spec.rb
@@ -1,20 +1,6 @@
 os_family = os.family
 
-describe user('test_user') do
-  it { should exist }
-  its('uid') { should eq 9001 } unless os_family == 'darwin'
-  case os_family
-  when 'suse'
-    its('groups') { should eq %w( users testgroup nfsgroup ) }
-  when 'darwin'
-    its('groups') { should include 'testgroup' }
-    its('groups') { should include 'nfsgroup' }
-  else
-    its('groups') { should eq %w( test_user testgroup nfsgroup ) }
-  end
-  its('shell') { should eq '/bin/bash' }
-end
-
+# Test if the groups exist
 describe group('testgroup') do
   it { should exist }
   its('gid') { should eq 3000 }
@@ -25,17 +11,122 @@ describe group('nfsgroup') do
   its('gid') { should eq 4000 }
 end
 
-describe user('test_user_keys_from_url') do
+# Test users from attributes
+describe user('usertoremove') do
+  it { should_not exist }
+end
+
+describe user('user_with_dev_null_home') do
+  it { should exist }
+  its('uid') { should eq 5000 } unless os_family == 'darwin'
+  its('gid') { should eq 4000 } unless os_family == 'darwin'
+  case os_family
+  when 'suse'
+    its('groups') { should eq %w( nfsgroup ) }
+  when 'darwin'
+    its('groups') { should include 'nfsgroup' }
+  else
+    its('groups') { should eq %w( nfsgroup ) }
+  end
+  its('shell') { should eq '/bin/bash' }
+end
+
+describe user('user_with_nfs_home_first') do
+  it { should exist }
+  case os_family
+  when 'suse'
+    its('groups') { should eq %w( users nfsgroup ) }
+  when 'darwin'
+    its('groups') { should include 'nfsgroup' }
+  else
+    its('groups') { should eq %w( user_with_nfs_home_first nfsgroup ) }
+  end
+  its('shell') { should eq '/bin/sh' }
+end
+
+describe file('/home/user_with_nfs_home_first/.ssh/id_ed25519.pub') do
+  its('content') { should include('ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIC6aZDF+x28xIlZSgyfyh3IAkencLp1VCU7JXBhJcXNy cheftestuser@laptop') }
+end unless os_family == 'darwin' # InSpec runs as non-root and can't see these files
+
+describe user('user_with_nfs_home_second') do
+  it { should exist }
+  case os_family
+  when 'suse'
+    its('groups') { should eq %w( users nfsgroup ) }
+  when 'darwin'
+    its('groups') { should include 'nfsgroup' }
+  else
+    its('groups') { should eq %w( user_with_nfs_home_second nfsgroup ) }
+  end
+  its('shell') { should eq '/bin/sh' }
+end
+
+describe file('/home/user_with_nfs_home_second/.ssh/id_ecdsa') do
+  its('content') { should include("-----BEGIN OPENSSH PRIVATE KEY-----\nb3BlbnNzaC1rZXktdjEAAAAABG5vbmUAAAAEbm9uZQAAAAAAAAABAAAAaAAAABNlY2RzYS\n1zaGEyLW5pc3RwMjU2AAAACG5pc3RwMjU2AAAAQQQns8Ec3poQBm6r7zv/UZojvXjrUZVB\n59R4LzOBw8cS/2xSQrVH8qm2X8kB1y6nuyydK0bbQF1pnES1P+uvG6e9AAAAsD2Nf449jX\n+OAAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBCezwRzemhAGbqvv\nO/9RmiO9eOtRlUHn1HgvM4HDxxL/bFJCtUfyqbZfyQHXLqe7LJ0rRttAXWmcRLU/668bp7\n0AAAAgJp/B6o2OADM0+NlkgH1dFcOLK64jhr3ScbWK4iyRdOcAAAAVZm11bGxlckBzYnBs\ndGMxbWxsdmRsAQID\n-----END OPENSSH PRIVATE KEY-----\n") }
+end unless os_family == 'darwin' # InSpec runs as non-root and can't see these files
+
+describe file('/home/user_with_nfs_home_second/.ssh/id_ecdsa.pub') do
+  its('content') { should include('ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBCezwRzemhAGbqvvO/9RmiO9eOtRlUHn1HgvM4HDxxL/bFJCtUfyqbZfyQHXLqe7LJ0rRttAXWmcRLU/668bp70=') }
+end unless os_family == 'darwin' # InSpec runs as non-root and can't see these files
+
+describe user('user_with_local_home') do
+  it { should exist }
+  case os_family
+  when 'suse'
+    its('groups') { should eq %w( users nfsgroup ) }
+  when 'darwin'
+    its('groups') { should include 'nfsgroup' }
+  else
+    its('groups') { should eq %w( user_with_local_home nfsgroup ) }
+  end
+  its('shell') { should eq '/bin/sh' }
+end
+
+describe directory('/home/user_with_local_home') do
+  it { should exist }
+end unless os_family == 'darwin' # InSpec runs as non-root and can't see these files
+
+describe file('/home/user_with_local_home/.ssh/id_rsa') do
+  its('content') { should include("-----BEGIN OPENSSH PRIVATE KEY-----\nb3BlbnNzaC1rZXktdjEAAAAABG5vbmUAAAAEbm9uZQAAAAAAAAABAAAAMwAAAAtzc2gtZW\nQyNTUxOQAAACAummQxfsdvMSJWUoMn8odyAJHp3C6dVQlOyVwYSXFzcgAAAJjzcJxA83Cc\nQAAAAAtzc2gtZWQyNTUxOQAAACAummQxfsdvMSJWUoMn8odyAJHp3C6dVQlOyVwYSXFzcg\nAAAEC7TGfA0MU0mh0V39qw5RSThUo0idTtU2vCe9bJrHmyFS6aZDF+x28xIlZSgyfyh3IA\nkencLp1VCU7JXBhJcXNyAAAAFWZtdWxsZXJAc2JwbHRjMW1sbHZkbA==\n-----END OPENSSH PRIVATE KEY-----\n") }
+end unless os_family == 'darwin' # InSpec runs as non-root and can't see these files
+
+# Test users from databags
+describe user('databag_mwaddams') do
+  it { should_not exist }
+end
+
+describe directory('/home/databag_mwaddams') do
+  it { should_not exist }
+end
+
+describe group('databag_mwaddams') do
+  it { should_not exist }
+end
+
+describe user('databag_test_user') do
+  it { should exist }
+  its('uid') { should eq 9001 } unless os_family == 'darwin'
+  case os_family
+  when 'suse'
+    its('groups') { should eq %w( users testgroup ) }
+  when 'darwin'
+    its('groups') { should include 'testgroup' }
+  else
+    its('groups') { should eq %w( databag_test_user testgroup ) }
+  end
+  its('shell') { should eq '/bin/bash' }
+end
+
+describe user('databag_test_user_keys_from_url') do
   it { should exist }
   its('uid') { should eq 9002 } unless os_family == 'darwin'
   case os_family
   when 'suse'
-    its('groups') { should eq %w( users testgroup nfsgroup ) }
+    its('groups') { should eq %w( users testgroup ) }
   when 'darwin'
     its('groups') { should include 'testgroup' }
-    its('groups') { should include 'nfsgroup' }
   else
-    its('groups') { should eq %w( test_user_keys_from_url testgroup nfsgroup ) }
+    its('groups') { should eq %w( databag_test_user_keys_from_url testgroup ) }
   end
   its('shell') { should eq '/bin/bash' }
 end
@@ -47,7 +138,7 @@ ssh_keys = [
   'ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQCy3cbPekJYHAIa8J1fOr2iIqpx/7pl4giJYAG7HCfsunRRUq3dY1KhVw1BlmMGIDzNwcuNVIfBS5HS/wREqbHMXxbwAjWrMwUofWd09CTuKJZiyTLUC5pSQWKDXZrefH/Fwd7s+YKk1s78b49zkyDcHSnKxjN+5veinzeU+vaUF9duFAJ9OsL7kTDEzOUU0zJdSdUV0hH1lnljnvk8kXHLFl9sKS3iM2LRqW4B6wOc2RbXUnx+jwNaBsq1zd73F2q3Ta7GXdtW/q4oDYl3s72oW4ySL6TZfpLCiv/7txHicZiY1eqc591CON0k/Rh7eR7XsphwkUstoUPQcBuLqQPA529zBigD7A8PBmeHISxL2qirWjR2+PrEGn1b0yu8IHHz9ZgliX83Q4WpjXvJ3REj2jfM8hiFRV3lA/ovjQrmLLV8WUAZ8updcLE5mbhZzIsC4U/HKIJS02zoggHGHZauClwwcdBtIJnJqtP803yKNPO2sDudTpvEi8GZ8n6jSXo/N8nBVId2LZa5YY/g/v5kH0akn+/E3jXhw4CICNW8yICpeJO8dGYMOp3Bs9/cRK8QYomXqgpoFlvkgzT2h4Ie6lyRgNv5QnUyAnW43O5FdBnPk/XZ3LA462VU3uOfr0AQtEJzPccpFC6OCFYWdGwZQA/r1EZQES0yRfJLpx+uZQ==',
 ]
 
-describe file('/home/test_user_keys_from_url/.ssh/authorized_keys') do
+describe file('/home/databag_test_user_keys_from_url/.ssh/authorized_keys') do
   ssh_keys.each do |key|
     its('content') { should include(key) }
   end