+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This file is used to list changes made in each version of the users cookbook.
 
 ## Unreleased
 
+- Removed hard dependency on data bags. See upgrading.md for details
+- Added per user no_user_group option to skip creating user group with the same name.
+- Cleaned up the manage resource
+- Added pubkey_type helper
+- Added more integration tests and updated unit tests to chefspec.
+
 ## 5.7.0 - *2021-03-08*
 
 - Add manage_home to the remove action of the manage resource